ufm: Fix writing incorrect offset if offset exceeds 0x100
Root cause:
It declares the parameter of addr to uint8_t in set_provision_data_in_flash API and the range of its value is from 0 to 255.
Ex: if users want to write the data at 0x12c offset, the data will be accidently wrote at 0x2c offset.

Solution:
Change to declare this parameter to uint32_t and removes redundant code.

+++ apps/aspeed-pfr/src/Smbus_mailbox/Smbus_mailbox.h
@@ -223,7 +223,7 @@ typedef struct _PFM_STRUCTURE {
 
 #pragma pack()
 
-unsigned char set_provision_data_in_flash(uint8_t addr, uint8_t *DataBuffer, uint8_t DataSize);
+unsigned char set_provision_data_in_flash(uint32_t addr, uint8_t *DataBuffer, uint8_t DataSize);
 int get_provision_data_in_flash(uint32_t addr, uint8_t *DataBuffer, uint32_t length);
 // void ReadFullUFM(uint32_t UfmId,uint32_t UfmLocation,uint8_t *DataBuffer, uint16_t DataSize);
 unsigned char erase_provision_data_in_flash(void);
